[filament_view] Fixed light intensity type

+++ plugins/filament_view/filament_view_plugin.h
@@ -125,7 +125,7 @@ class FilamentViewPlugin : public flutter::Plugin, public FilamentViewApi, publi
     std::optional<FlutterError> ChangeLightColorByGUID(
       const int64_t guid,
       const std::string& color,
-      int64_t intensity
+      double intensity
     ) override;
     std::optional<FlutterError> EnqueueAnimation(const int64_t guid, int64_t animation_index)
       override;
